refactor: simplify TV focus handling and consolidate focus helpers into presentation package
