+++ app/src/main/java/com/rdwatch/androidtv/ui/details/models/DetailUiState.kt
@@ -160,3 +160,4 @@ data class DetailFocusState(
 enum class NavigationDirection {
     UP, DOWN, LEFT, RIGHT
 }
+
